Re-implemented Phrase to work well with the editor.
Because the editor treats the phrase-creation UI as a grid, that meant
making the way a Phrase is represented more like a grid.

However, a Phrase still needs to interface with the soundcipher library,
which has its own representation of a musical score. So the Phrase class
now manages both the grid representation of a musical score and a
soundcipher representation of a musical score.