Controller-related changes.
Created "ArrowButtonView". Draws an arrow pointed left or right given a
button object.

Created arrow buttons in editor. Will be used to control phrase length.

Added getters and setters to Scrollbar class. Made other small changes.

Created scrollbar in editor. Will be used to control the portion of the
grid that is visible.