Replaced previous implementation for saving and loading phrases.
The previous implementation was based on serialization through Java's
Serializable interface. The problem with using this approach for saving
and loading is that it is very sensitive to change---whether that be a
different system with a different JVM or a slight rewrite of the code.
If any of these variables are different since the time a file was saved,
the file won't load properly.

The new implementation is just JSON, which is more reliable.