Side-stepped a ControlP5 bug that manifested in the Presenter.
When the screen size changed and a Controller was moved to a location
that didn't exist in the original screen size, there would be no way of
interacting with the Controller. I don't know why. My solution is to
destroy and reconstruct the Presenter's cp5 object when the screen size
changes.