Tuned the way new features are unlocked.
Made it so at the very start of playing the game, one icon in the
Presenter is already unlocked. That way, the player is never without
something to interact with.
Sped up the time it takes to unlock the change-screen buttons.
Changed the sequence in which the Editor's controllers unlock.

Also: Made it so the phrase that is generated at the very start of
playing the game is guarenteed to be a relatively short phrase. This is
to marginally reduce the amount of complexity that is thrown at the
player at the start of the game. A shorter phrase is a less complex
phrase.