Fixed saving problem.
Since changing the renderer to P2D the exit() method wasn't being called
on program exit. Apparently, this method does not necessarily execute
when the program exits. However, another method, exitActual() does seem
to necessarily execute when the program exits so that is what I'm using
now. This gives the program the opportunity to write some information to
a save file.