Disabled fixed camera mode.
As much as I like it for the sake of conceptual completeness, it's
performance wasn't reliable enough, even with the P2D renderer. And I
had to make the animation updating a bit choppier in order to make it
better in sync with the music (last commit), which makes high speed
animation even less feasible. If choosing to having a reasonable fixed
camera mode comes at the expense of synchronization accuracy, it's an
easy choice to make.