Put a band-aid on a memory leak.
Still don't understand the source of the problem, but the problem is
much less significant now. The main contributor to memory was a
ControlP5 class called Textfield which had as a member variable a
PGraphics buffer. After making sure to nullify the variable, the heap
size of the program grows much less rapidly.